Created Invoke-RTR
Created Invoke-RTR:

Custom command that will create an RTR session, issue a command and output the relevant stdout/stderr results.

Set Invoke-RTR as the default endpoint ('CustomRTR') for Invoke-Command so that users can supply a HostId to create a session, issue a command and output results (instead of creating a session, keeping the ID and then issuing a command to the session ID).

I will create an identical 'CustomBatchRTR' command to match, and corresponding commands for the Responder and Admin permission levels.