Update Split-Param.ps1
Bugfixes for Split-Param to ensure that requests that have both Body and Query values are getting the parameters passed appropriately.

Before, when using Edit-Detection, only the DetectionIds were being passed when the groups of parameters were being created. Now it's passing everything.

I also updated the query parameter split to match how it was being done with body identifiers, but I don't know that there's actually any endpoints that will have a series of identifiers in the query and body values at the same time...