Update Invoke-RTR.ps1
Updated arguments to only encapsulate in quotation marks when arguments does not contain a parameter. Previously, runscript parameters would be enclosed which would cause an error. This should make it so that commands that need things in quotes (like 'cat', when a space is present in the path) will work and commands that have a parameter will be ignored (because the user has to put those in quotes anyway).