Removed credential handling and SecureString use
Removed ImportCred/ExportCred functionality for security purposes.

Removed SecureString functionality for Secret handling since it provided a false sense of security (the securestrings are displayed as plaintext in the Windows Event Logs eventually anyway).

Changed $Global:Falcon to $Script:Falcon because the $Falcon variable won't need to be used to call methods that are no longer available.