Added Search-MalQueryHash
Search-MalQueryHash runs a YARA hunt for a specific sha256 value. This command will perform the search and wait for up to 30 seconds to output the result. If the search is not complete before the 30 seconds is up, the command will output the RequestId and Status field, indicating that the user needs to check status later.