Created Read-Meta
Added Read-Meta function, which removes "meta" details from a response, captures them in $Script:Meta for processing with Invoke-Loop, and outputs field names and values using Write-Verbose for troubleshooting purposes. This is a major step in obscuring "resources/combined/errors" from the user and displaying only the data (based on the count of results).