Bugfix for issue # 20
Removed default "created_timestamp" filter for Get-FalconVulnerability. It was
causing problems with the "-All" loop, since it was getting set each time the
loop made a request, rather than reusing the initial value. With it removed,
users will be forced to input a filter, but it's a much easier fix than trying
to carry that value through the loop.