v2.1.7
Relocated logging functions to the [ApiClient] object instead of having them within 'Write-Result'. Doing it this way avoids the 'holding' of various data in order to send the results to Humio. 'Write-Result' itself has also been updated to no longer require a 'Time' value, which was used for logging.

Modified the 'check time' for an active token from 15 seconds to 60 seconds to help prevent issues where a token expires during a long-running API call.