v2.1.7
Removed 'Library' commands, due to the release of support for 'Custom Scripts' in Falcon Fusion workflows.

The existing library scripts are going to be re-formatted to be compatible with workflows, so that saving the scripts in your local Falcon instance will be used as the delivery mechanism, instead of requiring PSFalcon for it to work.