Renamed and updated private functions
Renamed 'Add-Property' to 'Set-Property' and modified it to update when the property is present on the object being evaluated and add it to the object if it doesn't exist.

Updated 'Get-RtrResult' to collect results across different response types while reducing logic overhead.

Updated 'Invoke-FalconRtr' to display all populated output fields (including 'false' boolean values) based on the total output of all devices.

Begun work to update 'Invoke-FalconDeploy' to support both 'File' (formerly 'Path', but still has the 'Path' alias) and 'Archive'. This command requires additional updates before it will be fully functional.