Updated 'SensorTag' commands
Modified 'Add-FalconSensorTag' and 'Remove-FalconSensorTag' to include the uninstall token of the target device and while adding and removing sensor tags with 'CsSensorSettings.exe' on Windows sensor versions v6.42 and above.

Modified 'Get-FalconSensorTag' to return the 'FalconSensorTags' values listed in a devices API response if the target device is Windows sensor version 6.42 or above.