Update identity-protection.ps1
Changed 'Query' to 'String' (with alias 'query' for compatibility) and removed 'Mutation' as 'String' can be used for both.

Updated command to no longer manipulate GraphQL statement, and instead add 'Variables' with the relevant 'Cursor' variable that is defined within the statement when using the 'All' parameter. If this variable definition (or the required 'pageInfo') is not present, a warning message will be output and the '-All' loop will not be attempted.