Added 'Test-StringPattern' and renamed commands
Created 'Test-StringPattern' private function for validating values passed to parameters that accept [object[]] values which are ultimately filtered down to [string]. This function is used to output error messages when the proper 'id' values are not found in objects submitted from the pipeline.

Removed 'OnDemand' from command names 'Get-FalconOnDemandScan', 'Start-FalconOnDemandScan' and 'Stop-FalconOnDemandScan'.