Update psf-config.ps1
Continued to add `FileVantageRuleGroup` and `FileVantagePolicy` to `Import-FalconConfig`. Not entirely sure how to ensure that existing `HostGroup` items are assigned because neither `FileVantageRuleGroup` or `FileVantagePolicy` retain `HostGroup` names...

`FileVantagePolicy` and `FileVantageRuleGroup` are properly added if they don't exist, rule groups are assigned to the policy, and the policy is enabled. Still needs testing to ensure that modified rule groups and policies are updated.