Update psf-config.ps1
Removed `FirewallPolicy` from `ModifyDefault` options as it is not allowed to be modified.

Moved trimming of `FirewallRule` names after verification of whether or not rules exist for target `FirewallGroup`.

Changed how code related to (inactive) `FirewallRule` evaluation is commented out.

Added parenthesis around policy statement to ensure that policies are evaluated when they have an `id` AND when they have `prevention_settings` or `settings`, not when they have `id` OR `settings`.